Rebase changes made by hand
