+++ IPython/nbconvert/exporters/__init__.py
@@ -1,9 +1,9 @@
 from .export import *
 from .html import HTMLExporter
 from .slides import SlidesExporter
-from .exporter import TemplateExporter
+from .templateexporter import TemplateExporter
 from .latex import LatexExporter
 from .markdown import MarkdownExporter
 from .python import PythonExporter
 from .rst import RSTExporter
-from .baseexporter import BaseExporter
+from .exporter import Exporter
