+++ IPython/nbconvert/exporters/export.py
@@ -18,7 +18,7 @@ from functools import wraps
 from IPython.nbformat.v3.nbbase import NotebookNode
 from IPython.config import Config
 
-from .exporter import TemplateExporter
+from .templateexporter import TemplateExporter
 from .html import HTMLExporter
 from .slides import SlidesExporter
 from .latex import LatexExporter
