+++ IPython/nbconvert/exporters/html.py
@@ -19,7 +19,7 @@ from IPython.utils.traitlets import Unicode, List
 from IPython.nbconvert import preprocessors
 from IPython.config import Config
 
-from .exporter import TemplateExporter
+from .templateexporter import TemplateExporter
 
 #-----------------------------------------------------------------------------
 # Classes
