+++ IPython/nbconvert/filters/__init__.py
@@ -1,7 +1,7 @@
 from .ansi import *
+from .citation import *
 from .datatypefilter import *
 from .highlight import *
 from .latex import *
 from .markdown import *
 from .strings import *
-from .citation import *
